+++ src/app/domain/user/UserCache.ts
@@ -1,62 +1,65 @@
-import { cache as GlobalCache, DefaultCache } from '~/lib/util'
+import { DefaultCache } from '~/lib/util'
 import { User, UserRepository } from './'
 
-export class UserCache extends DefaultCache {
+export class UserCache extends DefaultCache
+{
 	/*
 	 * Cache storage key prefix
 	 */
 	protected prefix = 'user'
 
 	/*
-	 * The _client storage bucket this should act upon
+	 * Valkey database number
 	 */
-	protected bucket = 1
+	protected db = 1
 
 	/*
 	 * Constructor
 	 */
-	constructor() {
-		// Always use global cache client
-		super({ client: GlobalCache.client() })
+	constructor()
+	{
+		super()
 	}
 
 	/*
 	 * Fetch model from cache. Optionally, cache a fresh copy if it doesn't exist
 	 */
-	public async fetch(id: number, cache = true): Promise<User | undefined> {
+	public async fetch(id: number, shouldCache = true): Promise<User | undefined>
+	{
 		// Find in cache
 		const data = await this.getRaw(id)
 
 		// Found?
-		if(data) {
+		if(data)
+		{
 			return JSON.parse(data)
 		}
 
 		// Cache it now?
-		if(cache) {
-			// Find fresh source
-			const source = await this.getSource(id)
+		if( ! shouldCache) { return undefined }
 
-			// Found?
-			if(source) {
-				// Add to cache
-				await this.save(source)
+		// Find fresh source
+		const source = await this.getSource(id)
 
-				// Return
-				return JSON.parse(JSON.stringify(source))
-			}
-		}
+		// Not found?
+		if( ! source) { return undefined }
+
+		// Add to cache
+		await this.save(source)
+
+		return JSON.parse(JSON.stringify(source))
 	}
 
 	/*
 	 * Save model in cache by given id
 	 */
-	public async saveById(id: number): Promise<boolean> {
+	public async saveById(id: number): Promise<boolean>
+	{
 		// Fresh data
 		const user = await this.getSource(id)
 
 		// Not found?
-		if( ! user) return false
+		if( ! user) { return false }
 
 		// Cache
 		return await this.set(user.id, JSON.stringify(user))
@@ -65,19 +68,19 @@ export class UserCache extends DefaultCache {
 	/*
 	 * Save given model in cache
 	 */
-	public async save(data: User, refresh = false): Promise<boolean> {
+	public async save(data: User, refresh = false): Promise<boolean>
+	{
 		// Init
 		let user: User = data
 
 		// Refresh?
-		if(refresh) {
+		if(refresh)
+		{
 			const source = await this.getSource(data.id)
 
-			if(source) {
-				user = source
-			} else {
-				return false
-			}
+			if( ! source) { return false }
+
+			user = source
 		}
 
 		// Cache
@@ -87,7 +90,8 @@ export class UserCache extends DefaultCache {
 	/*
 	 * Get fresh model data from source
 	 */
-	public async getSource(id: number): Promise<User | undefined> {
+	public async getSource(id: number): Promise<User | undefined>
+	{
 		// Init repo
 		const repository = new UserRepository()