+++ src/config/index.ts
@@ -1,9 +1,11 @@
 import { api, ApiConfig } from './api'
 import { app, AppConfig } from './app'
-import { cache, CacheConfig } from './cache'
+import { cache } from './cache'
+import type { CacheConnectionOptions } from '~/lib/util/cache'
 import { db, DbConfig } from './database'
 import { docs, DocsConfig } from './docs'
 import { log, LogConfig } from './log'
+import { mail, MailConfig } from './mail'
 import { search, SearchConfig } from './search'
 import { schedule, ScheduleConfig } from './schedule'
 import { session, SessionConfig } from './session'
@@ -15,6 +17,7 @@ export const config: ConfigDictionary = {
 	db,
 	docs,
 	log,
+	mail,
 	search,
 	session,
 	schedule
@@ -23,10 +26,11 @@ export const config: ConfigDictionary = {
 interface ConfigDictionary {
 	api: ApiConfig
 	app: AppConfig
-	cache: CacheConfig
+	cache: CacheConnectionOptions
 	db: DbConfig
 	docs: DocsConfig
 	log: LogConfig
+	mail: MailConfig
 	search: SearchConfig
 	session: SessionConfig
 	schedule: ScheduleConfig