+++ src/lib/util/mail/index.ts
@@ -1,0 +1,3 @@
+export * from './mail'
+export * from './mail-service'
+export * from './types'